Use same utility functions for attendee and department imports While working on figuring out how to best import staff from prior years, I ended up doing a lot of cleanup of the attendee and department imports. This also fixes a bug where using the API reference page always returned a forbidden error.
